requirements.txt
# Bot Telegram
python-telegram-bot==20.7
python-telegram-bot[job-queue]==20.7

# Transcription Audio
openai-whisper==20231117
speechrecognition==3.10.0

# Traitement Audio
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# Analyse du ton et émotions
transformers==4.36.0
torch==2.1.1
nltk==3.8.1

# Utilitaires
python-dotenv==1.0.0
requests==2.31.0
pillow==10.1.0
numpy==1.24.3
pandas==2.0.3

# Support multilingue
langdetect==1.0.9

# Logging et monitoring
loguru==0.7.2
tqdm==4.66.1

# Formatage SRT
pysrt==1.1.2
